🗂️ Modularization of call expressions
