+++ BackEnd/Interpreter/Statements/Conditional_Jump/Switch_case.ts
@@ -9,7 +9,8 @@ import {
   RuntimeVal,
   StringVal,
 } from "../../../values.ts";
-import { evaluate, evaluate_return_statement } from "../../interpreter.ts";
+import { evaluate_return_statement } from "../../Functions/Functions.ts";
+import { evaluate } from "../../interpreter.ts";
 
 /**
  * Evaluates a switch statement.
