+++ BackEnd/Interpreter/Statements/Loop/For_loop.ts
@@ -14,7 +14,8 @@ import {
   NumberVal,
   RuntimeVal,
 } from "../../../values.ts";
-import { evaluate, evaluate_return_statement } from "../../interpreter.ts";
+import { evaluate_return_statement } from "../../Functions/Functions.ts";
+import { evaluate } from "../../interpreter.ts";
 
 /**
  * Evaluates a for loop statement.
