+++ BackEnd/Interpreter/Statements/Loop/While_statement.ts
@@ -8,7 +8,8 @@ import {
   NumberVal,
   RuntimeVal,
 } from "../../../values.ts";
-import { evaluate, evaluate_return_statement } from "../../interpreter.ts";
+import { evaluate_return_statement } from "../../Functions/Functions.ts";
+import { evaluate } from "../../interpreter.ts";
 
 /**
  * Evaluates a boolean while statement by repeatedly executing the body as long as the condition is true.
