+++ BackEnd/Interpreter/interpreter.ts
@@ -32,11 +32,7 @@ import Environment from "../Scope/environment.ts";
 
 import {
   BooleanVal,
-  FunctionVal,
-  MAKE_BOOL,
-  MAKE_FUNCTION,
   MAKE_NUll,
-  NativeFnVal,
   NumberVal,
   RuntimeVal,
   StringVal,
@@ -52,6 +48,11 @@ import {
   evaluate_member_expression,
 } from "./Expressions/Expressions.ts";
 import { evaluate_minus_expression } from "./Expressions/Unary_expression.ts";
+import {
+  evaluate_call_expression,
+  evaluate_function_definition,
+  evaluate_return_statement,
+} from "./Functions/Functions.ts";
 import { evaluate_break_statement } from "./Statements/Break.ts";
 import {
   evaluate_else_statement,
@@ -98,116 +99,6 @@ const evaluate_identifier = (
 };
 
 /**
- * Evaluates a function definition statement.
- * @param funcDef - The function definition statement.
- * @param env - The environment for variable lookup.
- * @returns The evaluated value of the function definition.
- */
-const evaluate_function_definition = (
-  funcDef: FunctionDefinition,
-  env: Environment,
-): RuntimeVal => {
-  // Create a new environment that captures the current environment as the closure
-  const functionEnv = new Environment(env);
-
-  // Create the function value and store it in the current environment
-  const funcValue = MAKE_FUNCTION(
-    funcDef.params.map((param) => param.name),
-    funcDef.body,
-    functionEnv, // Pass the closure environment to the function value
-  );
-
-  env.declareVar(funcDef.name, funcValue, false);
-  return funcValue;
-};
-
-/**
- * Evaluates a call expression by invoking a function with the provided arguments.
- * @param expr The call expression to evaluate.
- * @param env The environment in which the expression is evaluated.
- * @returns The result of the function call.
- * @throws Error if the value being called is not a function.
- */
-export const evaluate_call_expression = (
-  expr: CallExpr,
-  env: Environment,
-): RuntimeVal => {
-  const args = expr.args.map((arg) => evaluate(arg, env));
-  const fn = evaluate(expr.caller, env);
-
-  if (fn.type === "native-fn") {
-    const result = (fn as NativeFnVal).call(args, env);
-    return result;
-  } else if (fn.type === "function") {
-    const funcValue = fn as FunctionVal;
-
-    if (expr.args.length !== funcValue.params.length) {
-      let caller = expr.caller as Identifier;
-      throw `RunTimeError: Function "${caller.symbol}" expects ${funcValue.params.length} arguments, but ${expr.args.length} were provided.`;
-    }
-
-    // Create a new environment for the function call, inheriting the closure from the function definition
-    const functionEnv = new Environment(funcValue.closure);
-
-    // Map function parameters to their corresponding arguments
-    for (let i = 0; i < funcValue.params.length; i++) {
-      const paramName = funcValue.params[i];
-      const argValue = evaluate(expr.args[i], env);
-      functionEnv.declareVar(paramName, argValue, false);
-    }
-
-    functionEnv.declareVar("hasReturn", MAKE_BOOL(false), false);
-
-    // Execute the function body
-    let returnValue: RuntimeVal = MAKE_NUll();
-    for (const statement of funcValue.body) {
-      if (statement.kind === "ReturnStatement") {
-        const returnStmt = statement as ReturnStatement;
-        if (returnStmt.value) {
-          returnValue = evaluate(returnStmt.value, functionEnv);
-          return returnValue;
-        } else {
-          return MAKE_NUll();
-        }
-      } else {
-        let result = evaluate(statement, functionEnv);
-        let detectedReturn = functionEnv.lookupVar("hasReturn") as BooleanVal;
-        if (detectedReturn.value == true) {
-          returnValue = result;
-          return returnValue;
-        } else {
-          continue;
-        }
-      }
-    }
-    return returnValue;
-  } else {
-    throw `RunTimeError: Cannot call non-function ${JSON.stringify(fn)}`;
-  }
-};
-
-/**
- * Evaluates a return statement.
- * @param returnStmt - The return statement to evaluate.
- * @param env - The environment for variable lookup.
- * @returns The evaluated value of the return statement.
- */
-export const evaluate_return_statement = (
-  returnStmt: ReturnStatement,
-  env: Environment,
-): RuntimeVal => {
-  if (returnStmt.value === undefined) {
-    return MAKE_NUll();
-  } else {
-    let return_result = evaluate(returnStmt.value, env) as RuntimeVal;
-    if (return_result.type !== "null") {
-      return return_result;
-    } else {
-      return MAKE_NUll();
-    }
-  }
-};
-/**
  * Evaluates an AST node by handling different node kinds and calling the corresponding evaluation functions.
  * @param astNode - The AST node to evaluate.
  * @returns The runtime value of the evaluated AST node.
@@ -237,9 +128,6 @@ export function evaluate(astNode: Stmt, env: Environment): RuntimeVal {
 
     case "Identifier":
       return evaluate_identifier(astNode as Identifier, env);
-
-    case "CallExpr":
-      return evaluate_call_expression(astNode as CallExpr, env);
 
     case "AssignmentExpression":
       return evaluate_assignment_expression(
@@ -306,6 +194,9 @@ export function evaluate(astNode: Stmt, env: Environment): RuntimeVal {
 
     case "FunctionDefinition":
       return evaluate_function_definition(astNode as FunctionDefinition, env);
+
+    case "CallExpr":
+      return evaluate_call_expression(astNode as CallExpr, env);
 
     case "ReturnStatement":
       return evaluate_return_statement(astNode as ReturnStatement, env);
